Pattern Parser: implement Functor, Applicative, Monad
These are very rough implementations that no doubt can be cleaned up
once we understand how they work. (We know they work because they pass
the type checker, but we don't know why they work.)

It would a probably also make sense to introduce MonadFail here, either
making Either an instance of it (which, weirdly, it is not by default,
though I thought it used to be) or introducing our own equivalent that
might be both clearer to read and less handist.