Pattern: Re-implement `Applicative Parser` using `liftA2`
We also start documenting how Functor Parser and Applicative Parser work.

This is in part an exercise in understanding Applicative, but it also
turns out that liftA2 is a bit more clear about what's going for this
Functor than <*> is.

Our implementation of Monad does not use Applicative, instead just
re-implmenting sequencing from scratch (as we used to do back in the dark
ages when Monad was not an Applicative), so we add a temporary test for
pure and liftA2 until the testing of the Monad Parser is running
Applicative code as well.

For some reason we need to explicitly import (liftA2) from GHC.Base, per
our last commit; we'll fix this when we figure out what the problem is.