Parser: Reformat Modad Parser implementation and add a comment
Because Functor is more general than Applicative, which is in turn more
general than Monad, we can implement Applicative with Monad operations and
Functor with Applicative and/or Monad operations. We don't currently do
this becuase we (naturally) wrote the instance definitions in the opposite
order. (Looking back on it, this was perhaps not the correct approach.) So
we should try rewriting Applicative and Functor in terms of the Monad
operations.