BROKEN haskell/: Add Signature to parser state
We add Signature to the parser state, a Parser combinator to return the
signature from the state, and run that on the input (adding also `gobble`
to consume the remaining input we're not yet consuming properly). This
now produces an empty signature as output, which is one step closer to
producing the parsed signature as output.

We leave the test broken since the system is not yet doing anything useful
anyway.