BROKEN haskell/ ftest: confirm symbols is parsing up to the end of its line
We do this by merely checking that the next bit of the parse input is
the stuff immediately after the symbols line (including the newline
after the statement terminator). We also add a placeholder `symbols`
parser that just consumes exactly what the input is to confirm that
this test works.