haskell/ Pattern: Theory/Statement organization
We're not yet parsing the contents of the statements themselves, but we've
reworked the parser declarations to clarify our syntax: one theory (which
is curently just zero or more `symbol` lines) followed by zero or more
statements that will be checked within that theory.

Later this will be expanded to allow us multiple named theories, with
axioms too, and multiple sets of statements using various combinations of
the theories.

Detailed changes include:
- Introduce `keyword` as optional spaces followed by the keyword itself
- `statement` is now returning a sat, unsat or valid Statement (albeit one
  that's not fully parsed)
- A bunch of changes to `main` to clarify how we print things and
  print each statement on a separate line